Separating out dock classes.
