+++ common_widgets.py
@@ -1,0 +1,10 @@
+from PySide2.QtWidgets import QFrame
+
+
+
+
+class HorizontalLine(QFrame):
+    def __init__(self):
+        super().__init__()
+        self.setFrameShape(QFrame.HLine)
+        self.setFrameShadow(QFrame.Sunken)
